+++ packages/core/src/hid/index.ts
@@ -1,1 +1,2 @@
+export * from "./hid-gamepad";
 export * from "./xinput";